Added version which supports token based authentication

+++ graphrag_kb_server/test/test_jwt.py
@@ -4,7 +4,7 @@ import shutil
 
 from graphrag_kb_server.model.jwt_token import JWTTokenData, JWTToken
 from graphrag_kb_server.model.error import Error
-from graphrag_kb_server.service.jwt import (
+from graphrag_kb_server.service.jwt_service import (
     rename_to_folder,
     generate_token,
     decode_token,
@@ -15,7 +15,8 @@ from graphrag_kb_server.config import cfg
 def regenerate_test_tennant() -> Tuple[str, JWTTokenData]:
     folder_name = "test"
     folder_path = cfg.graphrag_root_dir_path / folder_name
-    shutil.rmtree(folder_path)
+    if folder_path.exists():
+        shutil.rmtree(folder_path)
     token_data = JWTTokenData(
         name=folder_name, email="test@test.com", time_delta_minutes=None
     )
